DPQL.derive {pda}	R Documentation

DPQL derive

Description

DPQL derive

Usage

DPQL.derive(ipdata,control,config)

Arguments

ipdata
individual participant data

control
pda control data

config
local site configuration

Details

This step calculated the intermediate aggregated data (XtWX, XtWY, and YtWY) for each site. May need to be iterated several times until prespecified rounds are met.

Value

list(SiX, SiXY, SiY, ni)

References

Chongliang Luo, et al. (2021) dPQL: a lossless distributed algorithm for generalized linear mixed model with application to privacy-preserving hospital profiling. medRxiv, doi: 10.1101/2021.05.03.21256561.
Chongliang Luo, et al. (2020) Lossless Distributed Linear Mixed Model with Application to Integration of Heterogeneous Healthcare Data. medRxiv, doi: 10.1101/2020.11.16.20230730.

[Package pda version 1.2-5 Index]