DPQL.estimate {pda}	R Documentation

PDA DPQL estimation

Description

PDA DPQL estimation

Usage

DPQL.estimate(ipdata=NULL,control,config)

Arguments

ipdata
no need

control
PDA control

config
cloud configuration

Details

DPQL estimation: (iterative) weighted DLMM using AD from all sites

Value

list(risk_factor, risk_factor_heterogeneity, bhat, sebhat, uhat, seuhat, Vhat)

References

Chongliang Luo, et al. (2021) dPQL: a lossless distributed algorithm for generalized linear mixed model with application to privacy-preserving hospital profiling. medRxiv, doi: 10.1101/2021.05.03.21256561.
Chongliang Luo, et al. (2020) Lossless Distributed Linear Mixed Model with Application to Integration of Heterogeneous Healthcare Data. medRxiv, doi: 10.1101/2020.11.16.20230730.

[Package pda version 1.2-5 Index]