DPQL.initialize {pda}	R Documentation

DPQL initialize

Description

DPQL initialize

Usage

DPQL.initialize(ipdata,control,config)

Arguments

ipdata
individual participant data

control
pda control data

config
local site configuration

Details

To initialize, fit glm at each individual site and send the estimated effect size and variances to the lead site. This step may be optional if we just use zero's as initial effect sizes to start the PQL algorithm.

Value

init

[Package pda version 1.2-5 Index]