myglmmPQL {pda}	R Documentation

A flexible version of MASS::glmmPQL

Description

A flexible version of MASS::glmmPQL

Usage

myglmmPQL(formula.glm, formula, offset=NULL, family, data, fixef.init = NULL, 
                 weights=NULL, REML=T, niter=10, verbose=T)

Arguments

formula.glm
formula used to fit glm for initial fixed effects

formula
formula used to fit iterative lmer in PQL algorithm

offset
glm offset term

family
glm family

data
glm data

fixef.init
initial fixed effects estimates, set to zeros if NULL

weights
glm weights

REML
lmer logical scalar - Should the estimates be chosen to optimize the REML criterion (as opposed to the log-likelihood)?

niter
glmmPQL maximum number of iterations.

verbose
glmmPQL logical: print out record of iterations?

Details

Use lme4::lmer instead of nlme::varFixed in PQL iteration to allow REML

Value

An object wiht the same format as lmer.

[Package pda version 1.2-5 Index]